2025-06-12 agora additional edge-case scenario handlings (#834)
* feat: add feature to end vesting schedule immediately when substantially overvested

* refactor: rename from "stop vesting schedule" to "let vesting schedule end"

* feat: add behaviour to delete and re-create a vesting schedule

This happens when a schedule needs to be updated but the original schedule is claimable, unclaimed and end date passed. Those schedules cant be updated anymore, so the behavior is to delete those schedules and create new ones with the updated total amount. The amount from previous schedule is transferred as cliff.

* feat: gatekeep agora buttons more strongly based on round type

* feat: keep round type in URL

* feat: handle another edge-case of unclaimed overvested schedule

* feat: show cliff amount in actions list

* fix: improve handling of "vested so far"